+++ src/core/util.h
@@ -35,7 +35,7 @@
 #ifndef janet_exit
 #include <stdio.h>
 #define janet_exit(m) do { \
-    printf("C runtime error at line %d in file %s: %s\n",\
+    fprintf(stderr, "C runtime error at line %d in file %s: %s\n",\
         __LINE__,\
         __FILE__,\
         (m));\
@@ -50,7 +50,7 @@
 /* What to do when out of memory */
 #ifndef JANET_OUT_OF_MEMORY
 #include <stdio.h>
-#define JANET_OUT_OF_MEMORY do { printf("janet out of memory\n"); exit(1); } while (0)
+#define JANET_OUT_OF_MEMORY do { fprintf(stderr, "janet out of memory\n"); exit(1); } while (0)
 #endif
 
 /* Omit docstrings in some builds */
@@ -125,6 +125,11 @@ void janet_lib_inttypes(JanetTable *env);
 #endif
 #ifdef JANET_THREADS
 void janet_lib_thread(JanetTable *env);
+#endif
+#ifdef JANET_NET
+void janet_lib_net(JanetTable *env);
+void janet_net_deinit(void);
+void janet_net_markloop(void);
 #endif
 
 #endif